+++ pmd-core/src/main/java/net/sourceforge/pmd/rules/RuleFactory.java
@@ -260,7 +260,7 @@ public class RuleFactory {
                          .error(
                              message,
                              attrValue,
-                             language.getTerseName(),
+                             language.getId(),
                              supportedVersions
                          );
             }